typescript overhaul
