+++ package.json
@@ -1,61 +1,48 @@
 {
-  "name": "joy-con-webhid",
-  "version": "0.7.0",
-  "description": "Joy-Con over WebHID ",
-  "scripts": {
-    "start": "npx http-server",
-    "clean": "shx rm -rf ./dist && mkdir dist",
-    "lint": "npx eslint . --ext .js,.mjs --fix --ignore-pattern dist/ --ignore-pattern src/madgwick.js",
-    "build:index": "npx terser ./src/index.js --toplevel --comments /@license/ --ecma=8 -o ./dist/index.js",
-    "build:joycon": "npx terser ./src/joycon.js --toplevel --comments /@license/ --ecma=8 -o ./dist/joycon.js",
-    "build:parse": "npx terser ./src/parse.js --toplevel --comments /@license/ --ecma=8 -o ./dist/parse.js",
-    "build:madgwick": "npx terser ./src/madgwick.js --toplevel --comments /@license/ --ecma=8 -o ./dist/madgwick.js",
-    "build:connectRingCon": "npx terser ./src/connectRingCon.js --toplevel --comments /@license/ --ecma=8 -o ./dist/connectRingCon.js",
-    "build": "npm run clean && npm run build:index && npm run build:joycon && npm run build:parse && npm run build:madgwick && npm run build:connectRingCon",
-    "copyDeps": "shx cp ./node_modules/ahrs/Madgwick.js ./src/madgwick.js && shx sed -i 's/module.exports =/export/' ./src/madgwick.js",
-    "prepare": "npm run lint && npm run fix && npm run build && npm run copyDeps",
-    "fix": "npx prettier --write ."
-  },
-  "main": "./dist/index.js",
-  "browser": "./dist/index.js",
-  "exports": "./dist/index.js",
-  "module": "./dist/index.js",
-  "unpkg": "./dist/index.js",
-  "files": [
-    "demo",
-    "src",
-    "dist"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/tomayac/joy-con-webhid.git"
-  },
-  "author": "Thomas Steiner (tomac@google.com)",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/tomayac/joy-con-webhid/issues"
-  },
-  "homepage": "https://github.com/tomayac/joy-con-webhid#readme",
-  "devDependencies": {
-    "eslint": "^8.47.0",
-    "eslint-config-google": "^0.14.0",
-    "eslint-config-prettier": "^9.0.0",
-    "http-server": "^14.1.1",
-    "prettier": "^3.0.2",
-    "shx": "^0.3.4",
-    "terser": "^5.19.2"
-  },
-  "eslintConfig": {
-    "parserOptions": {
-      "ecmaVersion": 2020,
-      "sourceType": "module"
-    },
-    "extends": [
-      "google",
-      "prettier"
-    ]
-  },
-  "dependencies": {
-    "ahrs": "^1.3.2"
-  }
+	"name": "joy-con-webhid",
+	"version": "0.7.0",
+	"description": "Joy-Con over WebHID ",
+	"type": "module",
+	"keywords": [
+		"joycon",
+		"webhid",
+		"gamepad",
+		"controller"
+	],
+	"scripts": {
+		"start": "npx http-server",
+		"check": "npx @biomejs/biome check --write ./src/*.ts",
+		"build": "npx tsc && npx vite build",
+		"fix": "npx prettier --write ."
+	},
+	"exports": {
+		".": {
+			"import": "./dist/joy-con-webhid.es.js",
+			"types": "./dist/index.d.ts"
+		}
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/tomayac/joy-con-webhid.git"
+	},
+	"author": "Thomas Steiner (tomac@google.com)",
+	"license": "Apache-2.0",
+	"bugs": {
+		"url": "https://github.com/tomayac/joy-con-webhid/issues"
+	},
+	"homepage": "https://github.com/tomayac/joy-con-webhid#readme",
+	"devDependencies": {
+		"@biomejs/biome": "1.9.4",
+		"@types/node": "^24.0.1",
+		"@types/w3c-web-hid": "^1.0.6",
+		"http-server": "^14.1.1",
+		"i": "^0.3.7",
+		"npm": "^11.4.2",
+		"typescript": "^5.8.3",
+		"vite": "^6.3.5",
+		"vite-plugin-dts": "^4.5.4"
+	},
+	"dependencies": {
+		"ahrs": "^1.3.2"
+	}
 }
